@@ -13,13 +13,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^12.0.1",
-    "@angular/core": "^12.0.1",
-    "@angular/forms": "^12.0.1",
+    "@angular/common": "^13.1.1",
+    "@angular/core": "^13.1.1",
+    "@angular/forms": "^13.1.1",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^12.0.1",
-    "@angular/platform-browser-dynamic": "^12.0.1",
-    "@angular/router": "^12.0.1",
+    "@angular/platform-browser": "^13.1.1",
+    "@angular/platform-browser-dynamic": "^13.1.1",
+    "@angular/router": "^13.1.1",
     "@ionic-native/core": "^5.33.0",
     "@ionic-native/splash-screen": "^5.33.0",
     "@ionic-native/status-bar": "^5.33.0",
@@ -31,14 +31,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1200.1",
-    "@angular-devkit/build-angular": "~12.0.1",
-    "@angular-devkit/core": "~12.0.1",
-    "@angular-devkit/schematics": "~12.0.1",
-    "@angular/cli": "~12.0.1",
-    "@angular/compiler": "~12.0.1",
-    "@angular/compiler-cli": "~12.0.1",
-    "@angular/language-service": "~12.0.1",
+    "@angular-devkit/architect": "~0.1301.2",
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular-devkit/core": "~13.1.2",
+    "@angular-devkit/schematics": "~13.1.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler": "~13.1.1",
+    "@angular/compiler-cli": "~13.1.1",
+    "@angular/language-service": "~13.1.1",
     "@ionic/angular-toolkit": "~3.1.1",
     "@types/jasmine": "~3.7.4",
     "@types/jasminewd2": "~2.0.9",
@@ -54,7 +54,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~9.1.1",
     "tslint": "~6.1.3",
-    "typescript": "4.2.4"
+    "typescript": "4.5.4"
   },
   "description": "An Ionic project"
 }
